requirements.txt
mcp
fastapi
uvicorn[standard]
python-dateutil
docker
playwright
duckduckgo-search>=4.0.0
PyYAML
psutil
pytest
pytest-asyncio
paramiko
aiohttp
